src/distgeom.cpp: Fix compile problem if you are using Eigen 2 instead of Eigen 3. Does it still work with Eigen 3? Note that this does not fix the compile if you have neither.

+++ src/distgeom.cpp
@@ -37,7 +37,7 @@ namespace OpenBabel {
   public:
     DistanceGeometryPrivate(const unsigned int N)
     {
-      bounds = Eigen::MatrixXf(N, N);
+      bounds = Eigen::MatrixXf(static_cast<int>(N), static_cast<int>(N));
     }
     ~DistanceGeometryPrivate()
     { }
